+++ apps/meteor/client/views/room/Sidepanel/SidepanelItem/RoomSidepanelItem.tsx
@@ -1,27 +1,21 @@
-import type { IRoom, ISubscription } from '@rocket.chat/core-typings';
-import { useUserSubscription } from '@rocket.chat/ui-contexts';
+import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 import React, { memo } from 'react';
 
 import { goToRoomById } from '../../../../lib/utils/goToRoomById';
 import { useTemplateByViewMode } from '../../../../sidebarv2/hooks/useTemplateByViewMode';
 import { useItemData } from '../hooks/useItemData';
 
-export type RoomSidepanelItemProps = {
+type RoomSidepanelItemProps = {
 	openedRoom?: string;
-	room: IRoom;
+	room: SubscriptionWithRoom;
 	parentRid: string;
 	viewMode?: 'extended' | 'medium' | 'condensed';
 };
 
 const RoomSidepanelItem = ({ room, openedRoom, viewMode }: RoomSidepanelItemProps) => {
 	const SidepanelItem = useTemplateByViewMode();
-	const subscription = useUserSubscription(room._id);
 
-	const itemData = useItemData({ ...room, ...subscription } as ISubscription & IRoom, { viewMode, openedRoom }); // as any because of divergent and overlaping timestamp types in subs and room (type Date vs type string)
-
-	if (!subscription) {
-		return <SidepanelItem onClick={goToRoomById} is='a' {...itemData} />;
-	}
+	const itemData = useItemData(room, { viewMode, openedRoom });
 
 	return <SidepanelItem onClick={goToRoomById} {...itemData} />;
 };