+++ apps/meteor/server/models/FreeSwitchEvent.ts
@@ -1,0 +1,6 @@
+import { registerModel } from '@rocket.chat/models';
+
+import { db } from '../database/utils';
+import { FreeSwitchEventRaw } from './raw/FreeSwitchEvent';
+
+registerModel('IFreeSwitchEventModel', new FreeSwitchEventRaw(db));