+++ packages/apps-engine/deno-runtime/main.ts
@@ -22,6 +22,8 @@ import apiHandler from './handlers/api-handler.ts';
 import handleApp from './handlers/app/handler.ts';
 import handleScheduler from './handlers/scheduler-handler.ts';
 import registerErrorListeners from './error-handlers.ts';
+import { startMetricsReport } from "./lib/metricsCollector.ts";
+import { parseArgs } from "./lib/parseArgs.ts";
 
 type Handlers = {
     app: typeof handleApp;
@@ -127,6 +129,10 @@ async function main() {
     }
 }
 
+const mainArgs = parseArgs(Deno.args);
+
 registerErrorListeners();
 
 main();
+
+startMetricsReport(mainArgs.metricsReportFrequencyInMs);